finder.find() returns @typedef FoundItem
