+++ src/js/urlNotification/finder.js
@@ -1,6 +1,16 @@
 'use strict';
 
+const config = require('./config');
 const storage = require('./storage');
+
+/**
+ * @typedef {object} FoundItem
+ * @property {string} url
+ * @property {string} message
+ * @property {string} backgroundColor
+ * @property {string} fontColor
+ * @property {string} displayPosition
+ */
 
 /**
  * Find pattern for content script.
@@ -10,14 +20,14 @@ const storage = require('./storage');
  * - PatternItem.url matches url
  *
  * @param {string} url
- * @returns {(PatternItem|null)}
+ * @returns {(FoundItem|null)}
  */
 const find = function(url) {
   let i, len, patterns = storage.getAll();
 
   for (i = 0, len = patterns.length; i < len; i++) {
     if (patterns[i].status === 1 && makeRegExp(patterns[i].url).test(url)) {
-      return patterns[i];
+      return createData(patterns[i]);
     }
   }
 
@@ -44,6 +54,20 @@ const convertForMatching = function(url) {
     replace(/\*/g, function() {
       return '[0-9a-zA-Z-_]+';
     });
+};
+
+/**
+ * @param {PatternItem} item
+ * @returns {FoundItem}
+ */
+const createData = function(item) {
+  return {
+    url: item.url,
+    message: item.msg,
+    backgroundColor: item.backgroundColor,
+    fontColor: config.defaultFontColor(),
+    displayPosition: item.displayPosition,
+  };
 };
 
 module.exports.findFor = find;
